+++ history.md
@@ -5,6 +5,7 @@ History over changes
 - fixed particle error where some blue dots remained after Blizzard was completed
 - added worldState og worldEnergy til map-abandon-reset-config
 - more old greenscreen pictures cleaned up
+- added more mobs (icebear, bear, lion, sickrat, rat) - need animation tweak
 
 * V 1.73
 - added options to have 1x1 or 2x2 object sheets